Restructure plugin folders and files
